+++ app/src/main/java/woowacourse/payments/ui/common/transformation/GroupedVisualTransformation.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.transformation
+package woowacourse.payments.ui.common.transformation
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.TransformedText
